changed HeroBanner to .tsx and added more top padding
