+++ src/components/Banner/Banner.css
@@ -5,6 +5,7 @@
 }
 
 .hero {
+	padding-top: 10%;
 	height: 105vh;
 }
